uaccess: add va_is_user param to vm_map function
vm_map takes an output pointer parameter `va`, to which it outputs the
address of the newly mapped region. The issue here is that vm_map is
being called for both user and kernel, hence `va` may belong to either
user or kernel address space.

To ease the adoption of PAN, this patch introduces a new param
`va_is_user` to `vm_map` function, so that `va` can be written in a
proper manner (direct access or user-access function)

Signed-off-by: Seonghyun Park <seonghp@amazon.com>