ldelf: Add map_zi_and_cp_from_bin system call
Add a new system call that is semantially equivalent to composition of
map_zi and copy_from_ta_bin system calls. When compared to calling both
system calls sequentially, this new system call does not grant user
access until the RW segment is written. In fact, the memory need not be
granted to the user address space before the content is filled in.

The benefit of having this new system call is that we can avoid
allocating a bounce buffer to work around PAN since the buffer is not
yet user memory before setting the permission. As a result, if we
replace map_zi and copy_from_ta_bin with this single system call,
we can reduce the memory footprint, as we no longer need to create a
copy of the segment within the kernel address space.

Signed-off-by: Seonghyun Park <seonghp@amazon.com>